Fix trait dispatch, for loops, binary operator monomorphisation, etc
This "fix" makes the compiler several times slower and an order of
magnitude messier but at least it works for some definition of works.
Because the Vec lib is now working, this ends the long-standing
typeinference branch.  Main focus now needs to be shifted to large
cleanup efforts and fixing smaller bugs.